usage: Move RawUsageRecord dependency into api package
